Issue 291:	HtmlParseData should hold a unique list of URLs
Currently HtmlParseData holds a non-unique list of links in the page, meaning that if a URL appears in the page several times then it will appear in the list of links several times.

I can't think of a scenario where somebody will parse an html page and want the same link more than once.

We should hold a set instead of a list, thus having a unique list of links.