Feature: seed tracking
The function of this feature is to know which documents resulted from which
seed. This can be useful when crawling specific websites while registering
the websites with that website. In addition, it is useful to know when a seed
has been fully crawled according to the crawl configuration.

Modifications:
* Added administrating the DocID of the seed for all its offspring.
A database is added to to WorkQueues that couples each DocID with a count
indicating the number of offspring in the queue for this page. As a
result, inProcessPages is now non-optional and is even used in
non-resumable crawls as this process requires knowing which pages are
assigned to crawlers at any time.

* In addition, added a small patch to CrawlController that transfers the
crawl queue to a newly spawned thread in case of thread failure. The
reason is that these pages have been added to the inProcessPages queue
but will never be removed if a thread crashes because the list is then
dropped. This fix fixes this. Shouldn't happen either way, a crashed
thread indicates a bug of course. This change could potentially lead to
a repeatedly crashing thread if the cause is one of the retrieved pages,
however, this makes it easier to debug the issue itself.

* A new method, handleSeedEnd(int docid) is added to WebCrawler. This may
be overridden by subclasses to do something when a seed has finished
crawling.

* Another consequence of the inProcessPages database is that it is known whether any crawler is
processing a page or not. If inProcessPages is non-empty but any crawler
is not working, this is a bug that should be fixed. If WorkQueues is
non-empty but any crawler is not working, this is also a bug. Therefore,
it's safe to skip waiting. In addition, the timeout that waits before
cleaning up is now based on a join of the crawler threads instead of just
waiting 10 seconds. This is usually much faster and more safe.

* Store result of handleUrlBeforeProcess in a different variable to make
sure that setProcess is called on the untweaked URL. This also allows
handleUrlBeforeProcess to return null in order to skip crawling a page
after all.

* Added a runSync method to the frontier that allows some code to be run
while obtaining the Frontier mutex. Can be used to synchronize certain
activities between the crawlers. Useful for code interacting with the crawler software.

* Some changes on the use of WorkQueues - using only one method for get
and delete simultanously: shift. Much cleaner, and less code
duplication. Also using a mutex in put().

* Added copy constructor to WebURL and use it to make a copy during
processing of a URL to avoid changing it before marking it as finished.