Refactoring RobotstxtServer module. The RobotstxtServer approach did not match any specific "standard" of robotstxt, and also did not match what Google is using. Checking is a rule matches was done by checking if the user-agent matches or contains the part specified in the robots.txt file, or if the robots.txt file mentions a wildcard as user-agent. However, the general tendency seems to be that from the entire file of directives, just one set of directives is applied. This is the one with the most specific User-agent string. For example, if robots.txt lists "Googlebot" and "Googlebot-news", then Googlebot-news will only obey the rules in the Googlebot-news section. If that section is absent it will obey the "Googlebot" section.
The new RobotstxtParser now uses this same approach.

Additionally, path parsing did not work correctly. Major crawlers allow
the use of wildcards mid-path. For example, /product/*/shops.html would
block the shop-page of all products. The approach in the old
RobotstxtParser did not allow for this syntax.

A third issue that has been resolved is that when robots.txt was
requested for a HTTPS host, it would still request robots.txt from the
same host using HTTP, while this should of course be HTTPS. In addition,
redirects would cause the RobotstxtServer to give up. Now, it will allow
up to three redirects before giving up.