The canonicalization of request URLs alphabetizes the parameters
The parameters do not retain their original ordering in the URLs, which leads to time out
requests when the new canonicalzed URls are visited. The non existing location contains alphabetized parameters.

I am under the impression that the original author used a SortedMap with the assumption
it retains the ordering of the Entrys inserted. It, however, is using the natural comparitor
for sorting a string (alphabetization).

I have changed all instances of SortedMap during the url parameter conocalization
process with a LinkedHashMap, which retains the ordering of Entrys inserted.