Attempt to use already closed Environment object when shutting down crawl controller explicitly
When shutting down the Crawl Controller explicitly, sometimes the Berkeley DB environment is closed before the call to store the currently processed page URL. This causes an exception to be thrown by the BDB Environment:

```
java.lang.IllegalStateException: Attempt to use non-open Environment object().
        at com.sleepycat.je.Environment.checkHandleIsValid(Environment.java:2153) ~[je-5.0.73.jar:5.0.73]
        at com.sleepycat.je.Environment.beginTransactionInternal(Environment.java:1312) ~[je-5.0.73.jar:5.0.73]
        at com.sleepycat.je.Environment.beginTransaction(Environment.java:1284) ~[je-5.0.73.jar:5.0.73]
        at edu.uci.ics.crawler4j.frontier.InProcessPagesDB.removeURL(InProcessPagesDB.java:58) ~[crawler4j-4.1.jar:na]
        at edu.uci.ics.crawler4j.frontier.Frontier.setProcessed(Frontier.java:163) [crawler4j-4.1.jar:na]
        at edu.uci.ics.crawler4j.crawler.WebCrawler.run(WebCrawler.java:262) [crawler4j-4.1.jar:na]
        at java.lang.Thread.run(Thread.java:745) [na:1.7.0_79]

```

I’ve moved the check for the shutting down of the controller before we store currently processed page URL to the frontier. If the controller enters the shutting down state, then the Berkeley DB is already closed and an attempt to store/read anything from it will
result with exception.