Better memory management when reading the data from server and enforcing maxBytes
When reading the HttpEntity from the server in the cases when the server did not
respond with Content-Length, the read buffer was automatically allocated to the
size of maxBytes. The value of maxBytes is arbitrary and it usually is set to a
higher value (when setting it, we expect most of the content that we're interested in
to be less than the maximal download size), so for a lot of pages we end up
allocating more memory that is actually needed.
While most of the time this does not cause problems (when the maxDownloadSize is
a couple of MiB or even tens of MiB), when setting it to a higher value (in the
hundreds of MiB) we end up wasting a lot of memory per crawling thread.
In the cases when the download size is very high, it may even cause an out of
memory error (java.lang.OutOfMemoryError: Java heap space). This is how I noticed
it in the first place - I had a fairly large max download size (of about a GB) and
multiple crawler threads running (about 10).

As a solution, I'm avoiding allocating maxBytes in the cases when actual size of
the HttpEntity is not available and allocate a fairly small read buffer (of about
4KiB) and then read the InputStream up until maxBytes is reached (or there is no
more content available).
In the cases when the actual size of the HttpEntity is available, then the read
buffer is allocated to whichever is smaller - the actual size or maxBytes.