PathRules that timeout if regex takes too long
TimeoutablePathRule adds support to timeout Regexp that run for too long. You can configure them to consider that a timeout means a match, or a not match.

Personally I'd throw RegexpTimeoutException, but it may break some foreign subclasses, so I decided to stick to return false. The static version, matchesRobotsPattern throws RegexpTimeoutException if configured to fail on timeout since it will not break any existing code.